Trained agent

+++ p2_continuous-control/Critic.py
@@ -23,7 +23,7 @@ class Critic(nn.Module):
         super(Critic, self).__init__()
         self.seed = torch.manual_seed(seed)
         self.fcs1 = nn.Linear(state_size, fcs1_units)
-        self.fc2 = nn.Linear(fcs1_units+action_size, fc2_units)
+        self.fc2 = nn.Linear(fcs1_units + action_size, fc2_units)
         self.fc3 = nn.Linear(fc2_units, 1)
         self.reset_parameters()
 
@@ -35,6 +35,6 @@ class Critic(nn.Module):
     def forward(self, state, action):
         """Build a critic (value) network that maps (state, action) pairs -> Q-values."""
         xs = F.relu(self.fcs1(state))
-        x = torch.cat((xs, action.float()), dim=1)  # Should work with actor's output here.
+        x = torch.cat((xs, action), dim=1)  # Should work with actor's output here.
         x = F.relu(self.fc2(x))
         return self.fc3(x)
